EvolutionSimulator starts with skipable Selection.
New main loop of Simulator matches with steps in Evolution3WEB. Thus it
supports now 1ASAP with BACKGROUND runMode. Still not synchronized with
further actions in UI. Also, IDs do have a meaning and should be
considered in EvolutionSimulator.

Fixed bug in mini simulation, that uses timer instead of
simulationtimer.